+++ packages/stream_chat_flutter/lib/src/message_widget/message_widget.dart
@@ -1033,6 +1033,11 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
             }
           },
         ),
+      ...widget.customActions.map((e) => StreamChatContextMenuItem(
+            leading: e.leading,
+            title: e.title,
+            onClick: () => e.onTap?.call(widget.message),
+          ))
     ];
   }